Isolated ProjectBackend from ProjectsBackend
